added login logic and files

+++ controllers/api/usersRoutes.js
@@ -1,7 +1,7 @@
 //Ian
 const router = require('express').Router();
 const { users } = require('../../models');
-// const withAuth = require('../../utils/auth');
+const withAuth = require('../../utils/auth');
 
 // test get
 router.get('/', async (req, res) => {
@@ -13,7 +13,7 @@ router.get('/', async (req, res) => {
     }
 });
 
-router.post('/', async (req, res) => {
+router.post('/', withAuth, async (req, res) => {
     try {
         const newUser = await users.create(req.body);
 
@@ -23,9 +23,9 @@ router.post('/', async (req, res) => {
     }
 });
 
-router.put('/:id', /*withAuth,*/ async (req, res) => {
+router.put('/:id', withAuth, async (req, res) => {
     try {
-        // if (req.session.user_id === parseInt(req.params.id)) {
+        if (req.session.user_id === parseInt(req.params.id)) {
             const updateUser = await users.update(
                 {
                     ...req.body
@@ -42,18 +42,18 @@ router.put('/:id', /*withAuth,*/ async (req, res) => {
             }
 
             res.status(200).json(updateUser);
-        // } else {
-        //     res.status(401).json({ message: 'you are not this user, cannot update' });
-        //     return;
-        // }
+        } else {
+            res.status(401).json({ message: 'you are not this user, cannot update' });
+            return;
+        }
     } catch (err) {
         res.status(500).json(err);
     }
 });
 
-router.delete('/:id', /*withAuth,*/ async (req, res) => {
+router.delete('/:id', withAuth, async (req, res) => {
     try {
-        // if (req.session.user_id === parseInt(req.params.id)) {
+        if (req.session.user_id === parseInt(req.params.id)) {
             const deleteUser = await users.destroy({
                 where: {
                     id: req.params.id,
@@ -66,61 +66,61 @@ router.delete('/:id', /*withAuth,*/ async (req, res) => {
             }
 
             res.status(200).json(deleteUser);
-        // } else {
-        //     res.status(401).json({ message: 'you are not this user, cannot delete' });
-        //     return;
-        // }
+        } else {
+            res.status(401).json({ message: 'you are not this user, cannot delete' });
+            return;
+        }
     } catch (err) {
         res.status(500).json(err);
     }
 });
 
 //login routes
-// router.post('/login', async (req, res) => {
-//     try {
-//         // Find the user by email address, if none are found throws and error
-//         const userData = await user.findOne({ where: { email: req.body.email } });
+router.post('/login', async (req, res) => {
+    try {
+        // Find the user by email address, if none are found throws and error
+        const userData = await users.findOne({ where: { email: req.body.email } });
 
-//         if (!userData) {
-//             res
-//                 .status(400)
-//                 .json({ message: 'Incorrect email or password, please try again' });
-//             return;
-//         }
+        if (!userData) {
+            res
+                .status(400)
+                .json({ message: 'Incorrect email or password, please try again' });
+            return;
+        }
 
-//         //verifies password
-//         const validPassword = await userData.checkPassword(req.body.password);
+        //verifies password
+        const validPassword = await userData.checkPassword(req.body.password);
 
-//         if (!validPassword) {
-//             res
-//                 .status(400)
-//                 .json({ message: 'Incorrect email or password, please try again' });
-//             return;
-//         }
+        if (!validPassword) {
+            res
+                .status(400)
+                .json({ message: 'Incorrect email or password, please try again' });
+            return;
+        }
 
-//         //creates session data
-//         req.session.save(() => {
-//             req.session.user_id = userData.id;
-//             req.session.logged_in = true;
+        //creates session data
+        req.session.save(() => {
+            req.session.user_id = userData.id;
+            req.session.logged_in = true;
 
-//             res.json({ user: userData, message: 'You are now logged in!' });
-//         });
+            res.json({ user: userData, message: 'You are now logged in!' });
+        });
 
-//     } catch (err) {
-//         res.status(400).json(err);
-//     }
-// });
+    } catch (err) {
+        res.status(400).json(err);
+    }
+});
 
-// // logout route
-// router.post('/logout', (req, res) => {
-//     if (req.session.logged_in) {
-//         req.session.destroy(() => {
-//             res.status(204).end();
-//         });
-//     } else {
-//         res.status(404).end();
-//     }
-// });
+// logout route
+router.post('/logout', (req, res) => {
+    if (req.session.logged_in) {
+        req.session.destroy(() => {
+            res.status(204).end();
+        });
+    } else {
+        res.status(404).end();
+    }
+});
 
 
 module.exports = router;
